pyproject.toml
[tool.poetry]
name = "fcm-adapter"
version = "0.1.2"
description = "Firebase Cloud Messaging Adapter"
authors = []
license = "MIT"
readme = "README.md"
packages = [{include = "fcm_adapter"}]

[tool.poetry.dependencies]
python = "^3.8"
httpx = "^0.23.3"
pyjwt = "^2.8.0"
cryptography = "^43.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
black = "^22.12.0"

[tool.black]
line-length = 79
target-version = ["py38"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
